put common variables/functions in separate file
