+++ src/components/Explorer/RunResultChart.jsx
@@ -1,17 +1,27 @@
 import React from "react";
 import * as d3 from 'd3';
 
+import {
+    cvgCartoonMap,
+    genomeBins,
+    colorMap,
+    sectionMargin,
+    sectionWidth,
+    sectionHeight,
+    tableShiftY,
+    barWidth,
+    barHeight,
+    barBorder,
+    caretWidth,
+    drawLegend,
+    addHeaders,
+    addColumns
+} from './ResultChartHelpers';
+
 const chartId = "runResultChart"
 
 export default () => {
     return <div id={chartId} />
-}
-
-var cvgCartoonMap = {
-    '_': 0,
-    '.': 0.25,
-    'o': 0.5,
-    'O': 1
 }
 
 var colMap = {
@@ -50,24 +60,6 @@ var colMap = {
 }
 
 var columns;
-var cvgLength = 25;
-var genomeBins = [...Array(cvgLength).keys()];
-
-var colorMap = d3.scaleSequential(d3.interpolateYlOrRd)
-        .domain([0, 1]);
-
-var colorScale = Object.values(cvgCartoonMap).map((value) => colorMap(value));
-
-var sectionMargin = {top: 2, right: 230, bottom: 2, left: 200};
-var sectionWidth = 750;
-var sectionHeight = 20;
-var tableShiftY = 40;
-var barWidth = sectionWidth - sectionMargin.left - sectionMargin.right;
-var barHeight = sectionHeight - sectionMargin.top - sectionMargin.bottom;
-
-var barBorder = {size: 1, color: '#999'};
-var rowLabelShiftX = -25;
-
 var summary;
 var maxAccessions = 10;
 
@@ -86,7 +78,7 @@ export const renderChart = (summaryParam, columnsParam) => {
     var columnHeadersG = chartSvg.append("g")
         .attr("transform", `translate(0, ${tableShiftY - sectionHeight})`);
     addHeaders(columnHeadersG);
-    addColumns(columnHeadersG);
+    addColumns(columnHeadersG, columns, colMap);
 
     var familyTextHeight = 50;
 
@@ -104,7 +96,7 @@ export const renderChart = (summaryParam, columnsParam) => {
             .attr("class", "family")
             .attr("rowid", `${family.family}`);
         var familySubGroup = drawExpandableRow(familyG, family.family, "family", familyCoverageData, i);
-        addColumns(familyG.select("svg"), family);
+        addColumns(familyG.select("svg"), columns, colMap, family);
         addFamilyText(familySubGroup, family);
 
         var familyAccessionsG = familySubGroup.append("g")
@@ -125,22 +117,10 @@ export const renderChart = (summaryParam, columnsParam) => {
             var accessionSubGroup = drawExpandableRow(
                 accessionG, accession.acc, "accession",
                 accessionCoverageData, i);
-            addColumns(accessionG.select("svg"), accession);
+            addColumns(accessionG.select("svg"), columns, colMap, accession);
             addAccessionText(accessionSubGroup, family, accession);
         });
     });
-}
-
-function linspace(start, end, n) {
-    var out = [];
-    var delta = (end - start) / (n - 1);
-    var i = 0;
-    while(i < (n - 1)) {
-        out.push(start + (i * delta));
-        i++;
-    }
-    out.push(end);
-    return out;
 }
 
 function getFamilyCoverageData(family) {
@@ -165,147 +145,6 @@ function getAccessionCoverageData(accession) {
         })
     });
     return coverageData;
-}
-
-// adapted from https://bl.ocks.org/starcalibre/6cccfa843ed254aa0a0d
-function drawLegend(svgElement) {
-    var legendWidth = 80,
-        legendHeight = 200,
-        margin = {top: 10, right: 60, bottom: 10, left: 2};
-
-    var legendSvg = svgElement.append("svg")
-        .attr("width", legendWidth)
-        .attr("height", legendHeight);
-
-    var gradient = legendSvg.append('defs')
-        .append('linearGradient')
-        .attr('id', 'gradient')
-        .attr('x1', '0%') // bottom
-        .attr('y1', '100%')
-        .attr('x2', '0%') // to top
-        .attr('y2', '0%')
-        .attr('spreadMethod', 'pad');
-
-    var pct = linspace(0, 100, colorScale.length).map(function(d) {
-        return Math.round(d) + '%';
-    });
-
-    var colourPct = d3.zip(pct, colorScale);
-    colourPct.forEach(function(d) {
-        gradient.append('stop')
-            .attr('offset', d[0])
-            .attr('stop-color', d[1])
-            .attr('stop-opacity', 1);
-    });
-
-    legendSvg.append('rect')
-        .attr('x1', 0)
-        .attr('y1', 0)
-        .attr('width', legendWidth - margin.left - margin.right)
-        .attr('height', legendHeight - margin.top - margin.bottom)
-        .style('fill', 'url(#gradient)')
-        .attr("transform", `translate(1, ${margin.top})`)
-        .style("stroke", "black")
-        .style("stroke-width", 1);
-
-    // create a scale and axis for the legend
-    var legendScale = d3.scaleLinear()
-        .domain([0, 1])
-        .range([legendHeight - margin.top - margin.bottom, 0]);
-    var legendAxis = legendSvg.append("g")
-        .attr("transform", `translate(${legendWidth - margin.left - margin.right}, ${margin.top - 0.5})`)
-        .call(d3.axisRight(legendScale));
-};
-
-function addHeaders(gElement) {
-    var yShift = 15;
-
-    var colText = "Match";
-    var xShift = sectionMargin.left + rowLabelShiftX;
-    var textG = gElement.append("g")
-    var text = textG.append("text")
-        .text(colText)
-        .style("text-anchor", "end")
-        .attr("transform",
-              `translate(${xShift}, ${yShift})`);
-
-    colText = "Coverage Heatmap";
-    xShift = sectionMargin.left + (barWidth / 2);
-    text = textG.append("text")
-        .text(colText)
-        .style("text-anchor", "middle")
-        .attr("transform",
-              `translate(${xShift}, ${yShift})`);
-}
-
-function addColumns(gElement, summaryEntry=null) {
-    var yShift = 15;
-    var colHeight = sectionHeight;
-    var textG = gElement.append("g")
-        .attr("transform",
-              `translate(${sectionMargin.left + barWidth + 10}, ${yShift})`);
-    var prevWidth = 0;
-    columns.forEach((column) => {
-        var colAttrs = colMap[column];
-        var colWidth = colAttrs["size"];
-        var colText = summaryEntry ? summaryEntry[column] : colAttrs["name"];
-        var cellG = textG.append("g");
-        var text = textG.append("text")
-            .text(colText)
-            .style("text-anchor", "middle")
-            .attr("x", prevWidth + (colWidth / 2));
-        if (summaryEntry) {
-            if (colText != null) {
-                text.text(colText + colAttrs.valueSuffix)
-                    .style("opacity", 0)
-                    .attr("column", column)
-                    .style("font-size", 12);
-                var diff = parseInt(colText) - colAttrs["domain"][0];
-                var diffCapped = Math.min(diff, colAttrs["domain"][1]);
-                var range = colAttrs["domain"][1] - colAttrs["domain"][0];
-                var colorBarWidth = Math.max(0, colWidth * diffCapped / range);
-                var colorBar = cellG.append("rect")
-                    .attr("fill", colAttrs["fill"])
-                    .attr("width", colorBarWidth)
-                    .attr("height", colHeight)
-                    .attr("x", prevWidth)
-                    .attr("y", -yShift);
-                var border = cellG.append("rect")
-                    .attr("fill", "none")
-                    .style("stroke", "black")
-                    .style("stroke-width", 1)
-                    .attr("width", colWidth)
-                    .attr("height", colHeight)
-                    .attr("x", prevWidth)
-                    .attr("y", -yShift);
-            }
-        }
-        var tooltipFontSize = 10;
-        var tooltipX = sectionMargin.left + barWidth + prevWidth + (colWidth / 2);
-        var tooltipY = 15;
-        var hoverForColumnText = textG.append("rect")
-            .attr("width", colWidth)
-            .attr("height", colHeight)
-            .attr("x", prevWidth)
-            .attr("y", -yShift)
-            .style("opacity", 0)
-            .on("mouseover", () => {
-                d3.selectAll(`[column="${column}"]`).style("opacity", 1);
-                d3.select("#tooltip")
-                    .text(colAttrs["desc"])
-                    .attr("font-size", tooltipFontSize)
-                    .attr("x", tooltipX)
-                    .attr("y", tooltipY)
-                    .style("text-anchor", "middle")
-                    .style("opacity", 1);
-            })
-            .on("mouseout", () => {
-                d3.selectAll(`[column="${column}"]`).style("opacity", 0);
-                d3.select("#tooltip").style("opacity", 0);
-            });
-
-        prevWidth += colWidth;
-    });
 }
 
 function drawExpandableRow(gElement, name, dataBin, heatSquareData, rowIndex) {
@@ -337,7 +176,7 @@ function drawExpandableRow(gElement, name, dataBin, heatSquareData, rowIndex) {
     yAxis.select("path").remove();
     yAxis.select("line").remove();
     yAxis.selectAll("text")
-        .attr("x", rowLabelShiftX)
+        .attr("x", -caretWidth)
         .style("font-size", 12)
         .style("fill", "blue")
         .style('cursor', 'pointer')
@@ -355,7 +194,7 @@ function drawExpandableRow(gElement, name, dataBin, heatSquareData, rowIndex) {
                 .append("a")
                 .attr("xlink:href", link)
             linkA.append("rect")
-                .attr("x", rowLabelShiftX - textWidth)
+                .attr("x", caretWidth - textWidth)
                 .attr("y", -8)
                 .attr("width", textWidth)
                 .attr("height", textHeight)
